User friendly items (#13)
StoryID: 25 (User-friendly pick-ups)

There are 3 actors that have changed:

BP_ItemSpawner: the item spawner is a c++ class with a blueprint version. All the code is in c++. When you place an item spawner in the level you can change what it spawns in the details panel under Item Spawning. the spawning is done in the c++ and the respawning is a function that is called on item destruction.
BP_AbilityItem: the ability item has been changed only a little bit, there is an extra bp block to call the respawn function on pickup.
BP_WeaponItemDebug: the same changes as BP_AbilityItem.

(possible) TODO: integrate ItemSpawner into a single object instead of separate objects for WeaponCore and Abilities

* master merge

* items working

* Revert "Merge branch 'master' into UserFriendlyItems"

This reverts commit cc2120e7d2c33c597aeb9229b641fa85d36584d4, reversing
changes made to 39d5a8c371bdb08dc076fe132ea67be19afc24bc.

* Master merge fix

* text update

* Removed items not coupled to spawners from MainLevel

---------

Co-authored-by: SanzianaTudose <tudosesanziana@gmail.com>