Damage abilites and player animations (#17)
Changed PlayerCharacterController to fix a bug in the rotation, the player now properly looks at the mouse location

Added a new character, with animations and an animation controller blending a combination of new animations together to make the caracter animations look proper when not moving directly at the mouse.

Updated UntitledGameMode to spawn the new character.

* Add damage to abilities

Just basically copy pasted the code from the default bullet to the abilityobjectcomponent with some compatability changes.

* Setup for proper animations

Just need to add sideways animations. I think if rotations is implemented correctly then this current implementation will need no changes

* Added new character+proper animations

* Revert "Setup for proper animations"

This reverts commit b11b562ceaaadce1b44e64cf28fe4da788ebf199.

* Fixed rotation issues

Fixed a bug in the original rotation code not rotating the player. And had to change the math of the animation a bit with the newly added rotations

* Fix: Player picking up Items

* kept the old Player BP (for reference) in a new directory

* Fix constructor issue in GameMode class

---------

Co-authored-by: SanzianaTudose <tudosesanziana@gmail.com>